requirements.txt
asgiref==3.5.2
Django==3.2
django-modeltranslation==0.18.9
Pillow==9.4.0
pytz==2022.4
sqlparse==0.4.3
